+++ src/Updates/UpdateEpsilon.h
@@ -1,5 +1,5 @@
-extern void UpdateEpsilon(double *P, double *Epsilon, double *Fst,
-                          int *NumAlleles, double lambda);
+extern void UpdateEpsilon(float *P, float *Epsilon, float *Fst,
+                          int *NumAlleles, float lambda);
 /*------------------------------------------*/
-extern void UpdateEpsilonCL(CLDict *clDict,double *P,double *Epsilon, double *Fst,
-              int *NumAlleles, double lambda);
+extern void UpdateEpsilonCL(CLDict *clDict,float *P,float *Epsilon, float *Fst,
+              int *NumAlleles, float lambda);
